# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wealth Master API",
    "version": "0.8.0",
    "description": "Session authentication, mandatory production MFA, accounts, calculated cash balances, manual income/expenses/refunds, and same-currency paired transfers. Current net worth by native currency, including archived account cash balances. User-owned two-level income/spending categories and optional ledger assignment. Dated normal monthly/yearly limits, specific-period exceptions, native currency actuals and explicit exception copying. Spending distribution reports show expenses before refunds and exact net spending independently of limits."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/budgets": {
      "get": {
        "summary": "Read spending limits and overlap-safe currency summaries",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetReport"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "periodType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "MONTH",
                "YEAR"
              ]
            },
            "description": "Defaults to MONTH"
          },
          {
            "name": "periodStart",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Canonical first day; defaults to current business period"
          }
        ],
        "description": "Explicit period exceptions plus resolved comparisons and overlap-safe summaries. Monthly exception/default takes precedence over annual exception/default. Annual comparisons use YTD spending; year views otherwise accrue monthly limits through the current month. Dated normal revisions preserve earlier periods. One owner-scoped repeatable-read snapshot; no-store."
      },
      "post": {
        "summary": "Create an owned available spending-category limit",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fresh masked token from GET /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetInput"
              }
            }
          }
        }
      }
    },
    "/budgets/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "put": {
        "summary": "Change limit only; fixed category/currency/period",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fresh masked token from GET /auth/csrf."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\"[0-9]+\"$"
            },
            "description": "Quoted response version."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetLimit"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete limit; permanent category references remain",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fresh masked token from GET /auth/csrf."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\"[0-9]+\"$"
            },
            "description": "Quoted response version."
          }
        ]
      }
    },
    "/budgets/copy": {
      "post": {
        "summary": "Copy selected versioned limits to another period",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetCopyResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fresh masked token from GET /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCopy"
              }
            }
          }
        },
        "description": "One source period and same period type required. All ownership/versions checked atomically before writes. Preserve existing targets and skip unavailable branches with explicit reasons. Source IDs are deduplicated."
      }
    },
    "/budgets/{id}/activity": {
      "get": {
        "summary": "Supporting expenses/refunds in ledger activity order",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "description": "50 items ordered by transaction date descending, creation timestamp descending, ID; includes archived accounts/categories. Read-only repeatable snapshot."
      }
    },
...
    "/budget-settings": {
      "get": {
        "summary": "Read owner normal settings and business date; no-store repeatable-read",
        "responses": {
          "400": {
            "description": "Invalid amount, currency, canonical period or promotion scope"
          },
          "401": {
            "description": "Authenticated MFA session required"
          },
          "403": {
            "description": "CSRF rejected"
          },
          "404": {
            "description": "Missing or unowned category/exception"
          },
          "409": {
            "description": "Unavailable category branch"
          },
          "412": {
            "description": "Expected setting/exception version or absence changed; no writes committed"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetSettings"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "summary": "Save dated normal setting from current period onward; expected head reference or absence",
        "responses": {
          "400": {
            "description": "Invalid amount, currency, canonical period or promotion scope"
          },
          "401": {
            "description": "Authenticated MFA session required"
          },
          "403": {
            "description": "CSRF rejected"
          },
          "404": {
            "description": "Missing or unowned category/exception"
          },
          "409": {
            "description": "Unavailable category branch"
          },
          "412": {
            "description": "Expected setting/exception version or absence changed; no writes committed"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetSetting"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetSettingInput"
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/budgets/effective-limit": {
      "put": {
        "summary": "Save period exception or atomically promote current-period limit to normal",
        "responses": {
          "400": {
            "description": "Invalid amount, currency, canonical period or promotion scope"
          },
          "401": {
            "description": "Authenticated MFA session required"
          },
          "403": {
            "description": "CSRF rejected"
          },
          "404": {
            "description": "Missing or unowned category/exception"
          },
          "409": {
            "description": "Unavailable category branch"
          },
          "412": {
            "description": "Expected setting/exception version or absence changed; no writes committed"
          },
          "204": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectiveLimitInput"
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "summary": "Remove explicit exception and restore applicable default; permanent history retained",
        "responses": {
          "400": {
            "description": "Invalid amount, currency, canonical period or promotion scope"
          },
          "401": {
            "description": "Authenticated MFA session required"
          },
          "403": {
            "description": "CSRF rejected"
          },
          "404": {
            "description": "Missing or unowned category/exception"
          },
          "409": {
            "description": "Unavailable category branch"
          },
          "412": {
            "description": "Expected setting/exception version or absence changed; no writes committed"
          },
          "204": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectiveLimitInput"
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/budgets/monthly-breakdown": {
      "get": {
        "summary": "Read twelve monthly comparisons for an owned category/currency/year; no-store",
        "responses": {
          "400": {
            "description": "Invalid amount, currency, canonical period or promotion scope"
          },
          "401": {
            "description": "Authenticated MFA session required"
          },
          "403": {
            "description": "CSRF rejected"
          },
          "404": {
            "description": "Missing or unowned category/exception"
          },
          "409": {
            "description": "Unavailable category branch"
          },
          "412": {
            "description": "Expected setting/exception version or absence changed; no writes committed"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyBreakdown"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9999
            }
          }
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "schemas": {
...
      "ActivityPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerOperation"
            }
          },
          "page": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "BudgetInput": {
        "type": "object",
        "required": [
          "categoryId",
          "currency",
          "periodType",
          "periodStart",
          "limit"
        ],
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "description": "Valid ISO 4217 currency"
          },
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "limit": {
            "type": "string",
            "pattern": "^[0-9]{1,20}(\\.[0-9]{1,8})?$"
          }
        }
      },
      "BudgetLimit": {
        "type": "object",
        "required": [
          "limit"
        ],
        "properties": {
          "limit": {
            "type": "string",
            "pattern": "^[0-9]{1,20}(\\.[0-9]{1,8})?$"
          }
        }
      },
      "Budget": {
        "type": "object",
        "required": [
          "id",
          "categoryId",
          "categoryName",
          "parentId",
          "parentName",
          "currency",
          "periodType",
          "periodStart",
          "limit",
          "actual",
          "remaining",
          "percentage",
          "overBudget",
          "available",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryName": {
            "type": "string"
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "parentName": {
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "type": "string"
          },
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "limit": {
            "type": "string"
          },
          "actual": {
            "type": "string"
          },
          "remaining": {
            "type": "string"
          },
          "percentage": {
            "type": [
              "string",
              "null"
            ],
            "description": "Two decimal places, HALF_UP; null for zero limits"
          },
          "overBudget": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BudgetSummary": {
        "type": "object",
        "required": [
          "currency",
          "netSpending",
          "uncategorized",
          "unbudgeted"
        ],
        "properties": {
          "currency": {
            "type": "string"
          },
          "netSpending": {
            "type": "string"
          },
          "uncategorized": {
            "type": "string"
          },
          "unbudgeted": {
            "type": "string"
          }
        }
      },
      "BudgetReport": {
        "type": "object",
        "required": [
          "periodType",
          "periodStart",
          "items",
          "currencies"
        ],
        "properties": {
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Budget"
            }
          },
          "currencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetSummary"
            }
          },
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "comparisons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetComparison"
            }
          }
        }
      },
...
      "BudgetCopy": {
        "type": "object",
        "required": [
          "sources",
          "targetPeriodStart"
        ],
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetSource"
            },
            "minItems": 1
          },
          "targetPeriodStart": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "BudgetCopyResult": {
        "type": "object",
        "required": [
          "created",
          "skipped"
        ],
        "properties": {
          "created": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Budget"
            }
          },
          "skipped": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetSkip"
            }
          }
        }
      },
...
      "BudgetReference": {
        "type": "object",
        "required": [
          "id",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BudgetSetting": {
        "type": "object",
        "required": [
          "id",
          "categoryId",
          "currency",
          "mode",
          "limit",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "NONE",
              "MONTH",
              "YEAR"
            ]
          },
          "limit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BudgetSettings": {
        "type": "object",
        "required": [
          "businessDate",
          "items"
        ],
        "properties": {
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetSetting"
            }
          }
        }
      },
      "BudgetSettingInput": {
        "type": "object",
        "required": [
          "categoryId",
          "currency",
          "mode"
        ],
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "NONE",
              "MONTH",
              "YEAR"
            ]
          },
          "limit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
              },
              {
                "type": "null"
              }
            ]
          },
          "expected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetReference"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "EffectiveLimitInput": {
        "type": "object",
        "required": [
          "categoryId",
          "currency",
          "periodType",
          "periodStart",
          "scope"
        ],
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "limit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope": {
            "type": "string",
            "enum": [
              "PERIOD",
              "NORMAL"
            ]
          },
          "expectedSetting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "expectedOverride": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetReference"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "MonthlyBreakdown": {
        "type": "object",
        "required": [
          "businessDate",
          "setting",
          "items"
        ],
        "properties": {
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "setting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetSetting"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "periodStart",
                "accrued",
                "comparison"
              ],
              "properties": {
                "periodStart": {
                  "type": "string",
                  "format": "date"
                },
                "accrued": {
                  "type": "boolean"
                },
                "comparison": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BudgetComparison"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}